fix(canon): map Options API bridge diagnostics
